Refactor: remove "_proto" from "media_proto.py"
See #2136
